some changes Co-authored-by: Abhishek Sachdeva <Abhishek312s@users.noreply.github.com>

+++ src/components/formCard/FormCard.jsx
@@ -149,6 +149,9 @@ function FormCard({
     const deleteForm = (deleteFormId) => {
         const token = localStorage.getItem('token');
 
+        if (!window.confirm('Are you sure you want to delete the form?'))
+            return;
+
         // console.log(deleteFormId);
         axios({
             method: 'DELETE',
